+++ react-native-expo/factories/mockData/assets.ts
@@ -1,0 +1,24 @@
+const ASSETS = [
+  {
+    name: 'Ethereum',
+    symbol: 'ETH',
+  },
+  {
+    name: 'Penumbra',
+    symbol: 'UM',
+  },
+  {
+    name: 'Osmosis',
+    symbol: 'OSMO',
+  },
+  {
+    name: 'ATOM',
+    symbol: 'Cosmos',
+  },
+  {
+    name: 'USDC',
+    symbol: 'USD Coin',
+  },
+];
+
+export default ASSETS;
